+++ src/screens/Favorites.js
@@ -1,0 +1,9 @@
+import Background from "../components/Background";
+
+export default function FavouritesScreen() {
+    return (
+        <Background title="Favorites">
+
+        </Background>
+    );
+}
